Abstract hide dynamic map in static map feature
